use AuthProvider to handle redirecting for gateway pages
