--- src/nextapp/components/no-gateway-redirect/index.ts
@@ -1,1 +1,0 @@
-export { default } from './no-gateway-redirect';
