+++ src/nextapp/pages/manager/consumers/index.tsx
@@ -36,7 +36,6 @@ import GrantAccessDialog from '@/components/access-request/grant-access-dialog';
 import ConsumerFilters from '@/components/consumer-filters';
 import AccessRequestsList from '@/components/access-request/access-requests-list';
 import { useNamespaceBreadcrumbs } from '@/shared/hooks';
-import NoGatewayRedirect from '@/components/no-gateway-redirect';
 
 const sortDate = new Intl.DateTimeFormat('en-ca', { dateStyle: 'short' });
 
@@ -77,8 +76,6 @@ export const getServerSideProps: GetServerSideProps = async (context) => {
 const ConsumersPage: React.FC<
   InferGetServerSidePropsType<typeof getServerSideProps>
 > = ({ queryKey }) => {
-  // Redirect to My Gateways page if no gateway selected
-  NoGatewayRedirect();
 
   const toast = useToast();
   const breadcrumbs = useNamespaceBreadcrumbs([{ text: 'Consumers' }]);
