+++ src/nextapp/pages/manager/gateways/detail.tsx
@@ -60,7 +60,6 @@ import useCurrentNamespace from '@/shared/hooks/use-current-namespace';
 import { useGlobal } from '@/shared/services/global';
 import EditNamespaceDisplayName from '@/components/edit-display-name';
 import { useNamespaceBreadcrumbs } from '@/shared/hooks';
-import NoGatewayRedirect from '@/components/no-gateway-redirect';
 
 const actions = [
   {
@@ -120,9 +119,6 @@ const secondaryActions = [
 ];
 
 const NamespacesPage: React.FC = () => {
-  // Redirect to My Gateways page if no gateway selected
-  NoGatewayRedirect();
-
   const { user } = useAuth();
   const breadcrumbs = useNamespaceBreadcrumbs();
   const hasNamespace = !!user?.namespace;
